requirements.txt
# Basic packages
numpy==1.24.1
opencv-python==4.6.0.66
# QR code decoder
pyzbar==0.1.9
# Object detection model interpreter
tflite-runtime==2.11.0
# Interface with GPIO devices
pyserial==3.5
smbus2==0.4.2
RPi.GPIO==0.7.1
# Equation solver for load cells calibration
sympy==1.12
# Extra packages for LED control
adafruit-blinka==8.19.0
rpi_ws281x==5.0.0
adafruit-circuitpython-neopixel==6.3.9